% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{friedman1}
\alias{friedman1}
\title{friedman1 dataset.}
\format{A data frame with 1000 rows and 11 numeric variables. All variables were scaled to 0-1.
X1 to X10 are the input variables and Ytrue is the output variable.}
\usage{
friedman1
}
\description{
A dataset generated with \code{\link[tgp]{friedman.1.data}} from 'tgp' package.
The friedman1 data set is commonly used to test semi-supervised regression methods.
}
\keyword{datasets}
