% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{predict.ssr}
\alias{predict.ssr}
\title{Predictions From a Fitted ssr Object}
\usage{
\method{predict}{ssr}(object, newdata, ...)
}
\arguments{
\item{object}{fitted object of class ssr.}

\item{newdata}{data frame with the input variables from which the response varaible is to be predicted.}

\item{...}{additional arguments (not used)}
}
\value{
A numeric vector with the predictions for each row of the input data frame.
}
\description{
Returns a vector of predicted responses from the fitted ssr object.
}
\examples{

# See ?ssr for an example.

}
